% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{efileAnnotations}
\alias{efileAnnotations}
\title{Read in the annotations from a multiple files}
\usage{
efileAnnotations(fileName)
}
\arguments{
\item{fileName}{vector of filenames}
}
\value{
a data frame
}
\description{
Read in annotations from multiple ELAN files
}

