% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{efileTierList}
\alias{efileTierList}
\title{List all tiers in elan file(s)}
\usage{
efileTierList(fileName, inheritMissingAttr = TRUE)
}
\arguments{
\item{fileName}{vector of filenames}

\item{inheritMissingAttr}{inherit missing attributes from parent tier (default=\code{TRUE})}
}
\value{
a data frame
}
\description{
List all annotation tiers in single or multiple ELAN files
}
\seealso{
\code{\link{readTierList}}
}

