% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{readAnnotations}
\alias{readAnnotations}
\title{Read annotations from the ELAN tree}
\usage{
readAnnotations(doc, annType = "ALIGNABLE", tierId = NULL,
  dropNACols = TRUE)
}
\arguments{
\item{doc}{the ELAN tree}

\item{annType}{either \code{"ALIGNABLE"} for time-stamped annotations or \code{"REF"} for reference annotations}

\item{tierId}{tier to read annotations from; if \code{NULL} will read from all tiers (default)}

\item{dropNACols}{boolean; drop any columns in the resulting dataset where all values are NAs (default is \code{TRUE})}
}
\value{
data frame with tier attributes and annotations in the field \code{VALUE}
}
\description{
Read annotations from the ELAN tree
}

