% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{readTimeSlots}
\alias{readTimeSlots}
\title{Read in the time slots}
\usage{
readTimeSlots(doc)
}
\arguments{
\item{doc}{parsed XML tree (see \code{\link{elanTree}})}
}
\value{
data frame with timeslot IDs and values
}
\description{
Read in the time slots
}

