% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{elanTree}
\alias{elanTree}
\title{Define an ELAN annotation tree from file}
\usage{
elanTree(fname)
}
\arguments{
\item{fname}{Name of xml or eaf file to read from}
}
\value{
parsed xml tree
}
\description{
Define an ELAN annotation tree from file
}

