% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{readTierList}
\alias{readTierList}
\title{Read the list of tiers in the ELAN file}
\usage{
readTierList(doc, inheritMissingAttrs = TRUE)
}
\arguments{
\item{doc}{the xml parse tree for the ELAN file}

\item{inheritMissingAttrs}{inherit missing attributes from parent tier (default=\code{TRUE})}
}
\value{
a data frame, where the rows are tiers and columns are attributes
}
\description{
Read the list of tiers in the ELAN file
}

