% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{readTimeSlots}
\alias{readTimeSlots}
\title{Read in the time slots}
\usage{
readTimeSlots(doc)
}
\arguments{
\item{doc}{parsed XML tree (see \code{\link{elanTree}})}
}
\value{
data frame with timeslot IDs and values
}
\description{
Read in the time slots
}

